remove cyclic deps
