+++ packages/shared/meta/sagas.ts
@@ -10,7 +10,7 @@ import {
 import { META_CONFIGURATION_INITIALIZED, metaConfigurationInitialized } from './actions'
 import defaultLogger from '../logger'
 import { FeatureFlagsName, MetaConfiguration, WorldConfig } from './types'
-import { getFeatureFlagEnabled, getFeatureFlagVariantValue, isMetaConfigurationInitiazed } from './selectors'
+import { getFeatureFlagVariantValue, isMetaConfigurationInitiazed } from './selectors'
 import { getSelectedNetwork } from 'shared/dao/selectors'
 import { SELECT_NETWORK } from 'shared/dao/actions'
 import { RootState } from 'shared/store/rootTypes'
@@ -19,8 +19,6 @@ import { trackEvent } from 'shared/analytics'
 import { addKernelPortableExperience } from 'shared/portableExperiences/actions'
 import { StorePortableExperience } from 'shared/types'
 import { getPortableExperienceFromUrn } from 'unity-interface/portableExperiencesUtils'
-import { waitForRendererInstance } from 'shared/renderer/sagas'
-import { getUnityInstance } from 'unity-interface/IUnityInterface'
 
 export function* waitForMetaConfigurationInitialization() {
   const configInitialized: boolean = yield select(isMetaConfigurationInitiazed)
@@ -62,18 +60,6 @@ function* initMeta() {
   yield put(metaConfigurationInitialized(merge))
 
   yield fork(fetchInitialPortableExperiences)
-  yield fork(configureFps)
-}
-
-// configures a cap on the FPS in the renderer
-function* configureFps() {
-  yield call(waitForRendererInstance)
-
-  const fpsCapped: boolean = yield select(getFeatureFlagEnabled, 'web_cap_fps')
-
-  if (fpsCapped) {
-    getUnityInstance().SetFpsTarget({ capped: fpsCapped, target: 30 })
-  }
 }
 
 function* fetchInitialPortableExperiences() {
