+++ packages/shared/renderer/sagas.ts
@@ -41,6 +41,8 @@ import { CommsContext } from 'shared/comms/context'
 import defaultLogger from 'shared/logger'
 import { receivePeerUserData } from 'shared/comms/peers'
 import { deepEqual } from 'atomicHelpers/deepEqual'
+import { ensureMetaConfigurationInitialized } from 'shared/meta'
+import { getFeatureFlagEnabled } from 'shared/meta/selectors'
 
 export function* waitForRendererInstance() {
   while (!(yield select(isRendererInitialized))) {
@@ -67,6 +69,19 @@ export function* rendererSaga() {
   yield call(listenToWhetherSceneSupportsVoiceChat)
 
   yield fork(reportRealmChangeToRenderer)
+  yield fork(configureFps)
+}
+
+// configures a cap on the FPS in the renderer
+function* configureFps() {
+  yield call(ensureMetaConfigurationInitialized)
+  yield call(waitForRendererInstance)
+
+  const fpsCapped: boolean = yield select(getFeatureFlagEnabled, 'web_cap_fps')
+
+  if (fpsCapped) {
+    getUnityInstance().SetFpsTarget({ capped: fpsCapped, target: 30 })
+  }
 }
 
 function* reportRealmChangeToRenderer() {
